refactor: Re-split FP FDM modules per issue #388 functional plan
Complete the module split following the original responsibility-based plan:

1. fp_fdm_bc.py (NEW) - Boundary condition enforcement
   - add_boundary_no_flux_entries (non-conservative)
   - add_boundary_no_flux_entries_conservative (flux FDM)

2. fp_fdm_advection.py (NEW) - Advection discretization
   - compute_advection_term_nd (upwind scheme)

3. fp_fdm_alg_flux.py (NEW) - Conservative flux discretization
   - add_interior_entries_conservative

4. fp_fdm_operators.py (REFACTORED) - Interior stencils only
   - is_boundary_point
   - add_interior_entries (non-conservative)

5. fp_fdm_time_stepping.py (REFACTORED) - Time integration only
   - Updated imports from responsibility-based modules

This eliminates the mixing of concerns from the initial split where
BC functions and conservative variants were bundled with operators.
