+++ mfg_pde/alg/numerical/fp_solvers/fp_fdm_time_stepping.py
@@ -3,10 +3,12 @@
 This module contains the forward time evolution logic for solving the FP equation
 using finite differences on multi-dimensional grids.
 
+Module structure per issue #388:
+    fp_fdm_time_stepping.py - When - time integration (forward Euler, implicit schemes)
+
 Functions:
     solve_fp_nd_full_system: Main time evolution loop for nD FP equation
     solve_timestep_tensor_explicit: Explicit timestep with tensor diffusion
-    compute_advection_term_nd: Advection term computation with upwind scheme
     solve_timestep_full_nd: Implicit timestep solver for scalar diffusion
 
 Mathematical Background:
@@ -36,11 +38,15 @@ import scipy.sparse as sparse
 from mfg_pde.geometry import BoundaryConditions
 from mfg_pde.utils.pde_coefficients import CoefficientField
 
-from .fp_fdm_operators import (
+# Import from responsibility-based modules per issue #388
+from .fp_fdm_advection import compute_advection_term_nd
+from .fp_fdm_alg_flux import add_interior_entries_conservative
+from .fp_fdm_bc import (
     add_boundary_no_flux_entries,
     add_boundary_no_flux_entries_conservative,
+)
+from .fp_fdm_operators import (
     add_interior_entries,
-    add_interior_entries_conservative,
     is_boundary_point,
 )
 
@@ -348,7 +354,7 @@ def solve_timestep_tensor_explicit(
     diffusion_term = divergence_tensor_diffusion_nd(M_current, Sigma, spacing, boundary_conditions)
 
     # Compute advection term: div(alpha * m)
-    # Use central differences for drift
+    # Use upwind scheme from fp_fdm_advection module
     advection_term = compute_advection_term_nd(
         M_current, U_current, coupling_coefficient, spacing, ndim, boundary_conditions
     )
@@ -357,96 +363,6 @@ def solve_timestep_tensor_explicit(
     M_next = M_current + dt * (diffusion_term - advection_term)
 
     return M_next
-
-
-def compute_advection_term_nd(
-    M: np.ndarray,
-    U: np.ndarray,
-    coupling_coefficient: float,
-    spacing: tuple[float, ...],
-    ndim: int,
-    boundary_conditions: Any,
-) -> np.ndarray:
-    """
-    Compute advection term div(alpha * m) where alpha = -coupling_coefficient * grad(U).
-
-    Uses upwind scheme for stability.
-
-    Parameters
-    ----------
-    M : np.ndarray
-        Density
-    U : np.ndarray
-        Value function
-    coupling_coefficient : float
-        Coupling coefficient
-    spacing : tuple
-        Grid spacing
-    ndim : int
-        Spatial dimension
-    boundary_conditions : BoundaryConditions
-        Boundary conditions
-
-    Returns
-    -------
-    np.ndarray
-        Advection term div(alpha * m)
-    """
-    # Compute drift: alpha = -coupling_coefficient * grad(U)
-    # Use upwind scheme for advection-dominated stability
-
-    advection = np.zeros_like(M)
-
-    if ndim == 2:
-        dx, dy = spacing
-
-        # Compute grad(U) with central differences for velocity field
-        grad_U_x = np.zeros_like(U)
-        grad_U_y = np.zeros_like(U)
-
-        # Central differences with boundary handling
-        grad_U_x[:, 1:-1] = (U[:, 2:] - U[:, :-2]) / (2 * dx)
-        grad_U_y[1:-1, :] = (U[2:, :] - U[:-2, :]) / (2 * dy)
-
-        # Drift velocity: alpha = -lambda * grad(U)
-        alpha_x = -coupling_coefficient * grad_U_x
-        alpha_y = -coupling_coefficient * grad_U_y
-
-        # Upwind scheme for advection: div(alpha * m)
-        # Use one-sided differences based on flow direction
-        # Positive velocity -> backward difference (upwind from left)
-        # Negative velocity -> forward difference (upwind from right)
-
-        # X-direction upwind flux divergence
-        # Forward difference: (flux[i+1] - flux[i]) / dx
-        # Backward difference: (flux[i] - flux[i-1]) / dx
-        flux_x = alpha_x * M
-
-        # Compute both forward and backward differences
-        d_flux_x_forward = np.zeros_like(M)
-        d_flux_x_backward = np.zeros_like(M)
-        d_flux_x_forward[:, :-1] = (flux_x[:, 1:] - flux_x[:, :-1]) / dx
-        d_flux_x_backward[:, 1:] = (flux_x[:, 1:] - flux_x[:, :-1]) / dx
-
-        # Select based on velocity direction (upwind)
-        advection += np.where(alpha_x >= 0, d_flux_x_backward, d_flux_x_forward)
-
-        # Y-direction upwind flux divergence
-        flux_y = alpha_y * M
-
-        d_flux_y_forward = np.zeros_like(M)
-        d_flux_y_backward = np.zeros_like(M)
-        d_flux_y_forward[:-1, :] = (flux_y[1:, :] - flux_y[:-1, :]) / dy
-        d_flux_y_backward[1:, :] = (flux_y[1:, :] - flux_y[:-1, :]) / dy
-
-        # Select based on velocity direction (upwind)
-        advection += np.where(alpha_y >= 0, d_flux_y_backward, d_flux_y_forward)
-
-    else:
-        # General nD (placeholder)
-        raise NotImplementedError(f"Advection term not yet implemented for {ndim}D")
-
-    return advection
 
 
 def solve_timestep_full_nd(
